requirements.txt
mcp[cli]
passlib
python-jose
jwt
requests
fastapi 
